requirements.txt
typedb-client
PyYAML